+++ src/api/request/userProfile.ts
@@ -1,0 +1,9 @@
+export interface IRequestPostUserProfile {
+	userId: string;
+	userName: string;
+	dept: string;
+	grade: number;
+	bio: string;
+}
+
+export type IRequestPatchUserProfile = IRequestPostUserProfile;
